complete attention_v0

+++ 7-attention/attention_vx.h
@@ -1,7 +1,7 @@
 #include <cuda_runtime.h>
 
 int attention_v0(const float *Q, const float *K, const float *V,
-                 const float *attnMasks,
+                 const int *attnMasks,
                  float *L, float *M, float *O,
                  int bs, int numHeads, int seqLen, int headDim,
                  cudaStream_t stream);
